Advanced-Programming-Techniques/AirTrend/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
biopython>=1.81
matplotlib>=3.7.0
pyyaml>=6.0
snakemake>=7.0.0
pytest>=7.0.0
requests>=2.31.0
urllib3>=2.0.0